Fixes #30. Add support for Shared Libraries
